Move Shields So It's Not a Route
